+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Holidays for these countries: Ireland, Russia.
+- Holidays for these countries: Brazil, Ireland, Russia.
+- Fat Tuesday in ChristianHolidays.
 - SortByNameFilter and SortByTypeFilter.
 
 ## [0.4.0] - 2020-11-02